rename RegisterEmail.js to 'RegisterEmailMutation.js
