+++ src/type/MutationType.js
@@ -1,17 +1,15 @@
 // @flow
 
-import {
-  GraphQLObjectType,
-} from 'graphql';
+import { GraphQLObjectType } from 'graphql';
 
 import LoginEmail from '../mutation/LoginEmailMutation';
-import RegisterEmail from '../mutation/RegisterEmail';
+import RegisterEmail from '../mutation/RegisterEmailMutation';
 import ChangePassword from '../mutation/ChangePasswordMutation';
 
 export default new GraphQLObjectType({
   name: 'Mutation',
   fields: () => ({
-    // auth
+		// auth
     LoginEmail,
     RegisterEmail,
     ChangePassword,
